requirements.txt
aiogram==3.8
asyncio
mysql-connector-python
python-dotenv